requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Flask-Login==0.6.3
Flask-Dance==7.1.0
blinker==1.7.0
python-dotenv==1.0.0
